requirements.txt
psutil==6.1.1
pycryptodomex==3.21.0
PyQt6==6.8.1
PyQt6_sip==13.10.0
python-dotenv==1.1.0
pywin32==308
